+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "PyIpsm"
-version = "0.0.1"
+version = "0.0.2"
 authors = [
   { name="Kurian Joseph Kattukaren", email="kuriankattukaren@gmail.com" },
 ]
